.github/actions/module_identifier/src/requirements.txt
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==42.0.2
Deprecated==1.2.14
idna==3.6
pycparser==2.21
PyGithub==2.2.0
PyJWT==2.8.0
PyNaCl==1.5.0
PyYAML==6.0.1
requests==2.31.0
typing_extensions==4.9.0
urllib3==2.2.0
wrapt==1.16.0
